fix: single-module
