+++ Makefile
@@ -14,77 +14,61 @@ MANIFEST_TAG=latest
 IMAGE_PREFIX=osint
 IMAGE_REGISTRY=local
 
-PHONY: all
+.PHONY: all
 all: build
 
-PHONY: build $(BUILD_TARGETS)
-build: go-test $(BUILD_TARGETS)
-%.build: %.go-test
+.PHONY: build
+build: $(BUILD_TARGETS)
+%.build: go-test
 	TARGET=$(*) IMAGE_TAG=$(IMAGE_TAG) IMAGE_PREFIX=$(IMAGE_PREFIX) BUILD_OPT="$(BUILD_OPT)" . hack/docker-build.sh
 
-PHONY: build-ci $(BUILD_CI_TARGETS)
+.PHONY: build-ci $(BUILD_CI_TARGETS)
 build-ci: $(BUILD_CI_TARGETS)
 %.build-ci:
 	TARGET=$(*) IMAGE_TAG=$(IMAGE_TAG) IMAGE_PREFIX=$(IMAGE_PREFIX) BUILD_OPT="$(BUILD_OPT)" . hack/docker-build.sh
 	docker tag $(IMAGE_PREFIX)/$(*):$(IMAGE_TAG) $(IMAGE_REGISTRY)/$(IMAGE_PREFIX)/$(*):$(IMAGE_TAG)
 
-PHONY: push-image $(IMAGE_PUSH_TARGETS)
+.PHONY: push-image $(IMAGE_PUSH_TARGETS)
 push-image: $(IMAGE_PUSH_TARGETS)
 %.push-image:
 	docker push $(IMAGE_REGISTRY)/$(IMAGE_PREFIX)/$(*):$(IMAGE_TAG)
 
-PHONY: pull-image $(IMAGE_PULL_TARGETS)
+.PHONY: pull-image $(IMAGE_PULL_TARGETS)
 pull-image: $(IMAGE_PULL_TARGETS)
 %.pull-image:
 	docker pull $(IMAGE_REGISTRY)/$(IMAGE_PREFIX)/$(*):$(IMAGE_TAG)
 
-PHONY: tag-image $(IMAGE_TAG_TARGETS)
+.PHONY: tag-image $(IMAGE_TAG_TARGETS)
 tag-image: $(IMAGE_TAG_TARGETS)
 %.tag-image:
 	docker tag $(SOURCE_IMAGE_PREFIX)/$(*):$(SOURCE_IMAGE_TAG) $(IMAGE_REGISTRY)/$(IMAGE_PREFIX)/$(*):$(IMAGE_TAG)
 
-PHONY: create-manifest $(MANIFEST_CREATE_TARGETS)
+.PHONY: create-manifest $(MANIFEST_CREATE_TARGETS)
 create-manifest: $(MANIFEST_CREATE_TARGETS)
 %.create-manifest:
 	docker manifest create $(IMAGE_REGISTRY)/$(IMAGE_PREFIX)/$(*):$(MANIFEST_TAG) $(IMAGE_REGISTRY)/$(IMAGE_PREFIX)/$(*):$(IMAGE_TAG_BASE)_linux_amd64 $(IMAGE_REGISTRY)/$(IMAGE_PREFIX)/$(*):$(IMAGE_TAG_BASE)_linux_arm64
 	docker manifest annotate --arch amd64 $(IMAGE_REGISTRY)/$(IMAGE_PREFIX)/$(*):$(MANIFEST_TAG) $(IMAGE_REGISTRY)/$(IMAGE_PREFIX)/$(*):$(IMAGE_TAG_BASE)_linux_amd64
 	docker manifest annotate --arch arm64 $(IMAGE_REGISTRY)/$(IMAGE_PREFIX)/$(*):$(MANIFEST_TAG) $(IMAGE_REGISTRY)/$(IMAGE_PREFIX)/$(*):$(IMAGE_TAG_BASE)_linux_arm64
 
-PHONY: push-manifest $(MANIFEST_PUSH_TARGETS)
+.PHONY: push-manifest $(MANIFEST_PUSH_TARGETS)
 push-manifest: $(MANIFEST_PUSH_TARGETS)
 %.push-manifest:
 	docker manifest push $(IMAGE_REGISTRY)/$(IMAGE_PREFIX)/$(*):$(MANIFEST_TAG)
 	docker manifest inspect $(IMAGE_REGISTRY)/$(IMAGE_PREFIX)/$(*):$(MANIFEST_TAG)
 
-PHONY: go-test $(TEST_TARGETS) pkg-test
-go-test: $(TEST_TARGETS) pkg-test
-%.go-test:
-	cd src/$(*) && GO111MODULE=on go test ./...
-pkg-test:
-	cd pkg/common && GO111MODULE=on go test ./...
+.PHONY: go-test
+go-test:
+	GO111MODULE=on go test ./...
 
-PHONY: go-mod-update
-go-mod-update:
-	cd src/subdomain \
-		&& go get -u \
-			github.com/ca-risken/core/... \
-			github.com/ca-risken/datasource-api/... \
-			github.com/ca-risken/osint/...
+.PHONY: lint
+lint:
+	GO111MODULE=on GOFLAGS=-buildvcs=false golangci-lint run --timeout 5m
 
-PHONY: go-mod-tidy
-go-mod-tidy:
-	cd pkg/common    && go mod tidy
-	cd src/subdomain && go mod tidy
-
-.PHONY: lint pkg-lint
-lint: $(LINT_TARGETS)  pkg-lint
-%.lint: FAKE
-	sh hack/golinter.sh src/$(*)
-pkg-lint:
-	sh hack/golinter.sh pkg/common
-
-.PHONY: workspace
-workspace:
-	go work use -r .
+.PHONY: enqueue-website
+enqueue-website:
+	aws sqs send-message \
+		--endpoint-url http://localhost:9324 \
+		--queue-url http://localhost:9324/queue/osint-website \
+		--message-body '{"osint_id":1002, "osint_data_source_id":1003,"rel_osint_data_source_id":1002, "data_source":"osint:website", "project_id":1001, "resource_name":"https://console.security-hub.jp", "resource_type":"website", "detect_word":""}'
 
 FAKE:
