+++ codebuild/multi-arch/buildspec-image.yml
@@ -4,17 +4,10 @@ env:
   variables:
     IMAGE_PREFIX: "risken-osint"
   parameter-store:
-    GITHUB_USER: "/build/GITHUB_USER"
-    GITHUB_TOKEN: "/build/GITHUB_TOKEN"
     DOCKER_USER: "/build/DOCKER_USER"
     DOCKER_TOKEN: "/build/DOCKER_TOKEN"
 
 phases:
-  install:
-    commands:
-      - echo "machine github.com" > ~/.netrc
-      - echo "login ${GITHUB_USER}" >> ~/.netrc
-      - echo "password ${GITHUB_TOKEN}" >> ~/.netrc
   pre_build:
     commands:
       - echo Setting environment variables
