+++ codebuild/multi-arch/buildspec-manifest.yml
@@ -3,16 +3,8 @@ version: 0.2
 env:
   variables:
     IMAGE_PREFIX: 'risken-osint'
-  parameter-store:
-    GITHUB_USER: '/build/GITHUB_USER'
-    GITHUB_TOKEN: '/build/GITHUB_TOKEN'
 
 phases:
-  install:
-    commands:
-      - echo "machine github.com" > ~/.netrc
-      - echo "login ${GITHUB_USER}" >> ~/.netrc
-      - echo "password ${GITHUB_TOKEN}" >> ~/.netrc
   pre_build:
     commands:
       - echo Setting environment variables
