+++ codebuild/release/buildspec-image.yml
@@ -5,16 +5,8 @@ env:
     PRIVATE_IMAGE_PREFIX: 'risken-osint'
     PUBLIC_REGISTRY: 'public.ecr.aws/risken'
     PUBLIC_IMAGE_PREFIX: 'osint'
-  parameter-store:
-    GITHUB_USER: '/build/GITHUB_USER'
-    GITHUB_TOKEN: '/build/GITHUB_TOKEN'
 
 phases:
-  install:
-    commands:
-      - echo "machine github.com" > ~/.netrc
-      - echo "login ${GITHUB_USER}" >> ~/.netrc
-      - echo "password ${GITHUB_TOKEN}" >> ~/.netrc
   pre_build:
     commands:
       - echo Setting environment variables
