+++ hack/docker-build.sh
@@ -5,4 +5,4 @@ fi
 if [ "${IMAGE_PREFIX}" = "" ]; then
   IMAGE_PREFIX=default_prefix
 fi
-cd src/${TARGET} && docker build ${BUILD_OPT} --build-arg GITHUB_USER=${GITHUB_USER} --build-arg GITHUB_TOKEN=${GITHUB_TOKEN} -t ${IMAGE_PREFIX}/${TARGET}:${IMAGE_TAG} .
+docker build ${BUILD_OPT} -t ${IMAGE_PREFIX}/${TARGET}:${IMAGE_TAG} -f dockers/${TARGET}/Dockerfile .
