--- pkg/common/go.mod
@@ -1,3 +1,0 @@
-module github.com/ca-risken/osint/pkg/common
-
-go 1.18
