--- src/subdomain/logger.go
@@ -1,5 +1,0 @@
-package main
-
-import "github.com/ca-risken/common/pkg/logging"
-
-var appLogger = logging.NewLogger()
