--- src/website/Dockerfile
@@ -1,34 +1,0 @@
-FROM golang:1.18.2 as builder
-WORKDIR /go/src/github.com/ca-risken/osint/src/website/
-ADD *.go go.* ./
-ARG GITHUB_USER
-ARG GITHUB_TOKEN
-RUN echo "machine github.com" > ~/.netrc
-RUN echo "login $GITHUB_USER" >> ~/.netrc
-RUN echo "password $GITHUB_TOKEN" >> ~/.netrc
-RUN go mod download
-RUN CGO_ENABLED=0 GOOS=linux go build -buildvcs=false -o /go/bin/
-
-FROM public.ecr.aws/risken/osint/website-base:v0.0.1
-COPY --from=builder /go/bin/website /usr/local/website/bin/
-ENV DEBUG= \
-  PROFILE_EXPORTER= \
-  PROFILE_TYPES= \
-  AWS_REGION= \
-  AWS_ACCESS_KEY_ID= \
-  AWS_SECRET_ACCESS_KEY= \
-  AWS_SESSION_TOKEN= \
-  SQS_ENDPOINT= \
-  WEBSITE_QUEUE_NAME= \
-  WEBSITE_QUEUE_URL= \
-  MAX_NUMBER_OF_MESSAGE= \
-  WAIT_TIME_SECOND=20 \
-  FINDING_SVC_ADDR= \
-  ALERT_SVC_ADDR= \
-  AWS_SVC_ADDR= \
-  RESULT_DIR=/tmp \
-  TZ=Asia/Tokyo
-WORKDIR /usr/local/website
-ENTRYPOINT ["/sbin/tini", "--", "/usr/local/bin/env-injector"]
-CMD ["bin/website"]
-
